Refactor file struture
