+++ slice/equals_test.go
@@ -4,38 +4,6 @@ import (
 	"fmt"
 )
 
-func ExampleContainsString() {
-	s := []string{"foo", "bar", "baz"}
-	fmt.Println(ContainsString(s, "foo"))
-	fmt.Println(ContainsString(s, "foobar"))
-	fmt.Println(ContainsString(nil, "foo"))
-	// Output:
-	// true
-	// false
-	// false
-}
-
-func ExampleContainsInt() {
-	s := []int{1, 2, 3}
-	fmt.Println(ContainsInt(s, 1))
-	fmt.Println(ContainsInt(s, 4))
-	fmt.Println(ContainsInt(nil, 4))
-	// Output:
-	// true
-	// false
-	// false
-}
-
-func ExampleUniqueStrings() {
-	fmt.Printf("%#v\n", UniqueStrings([]string{"foo", "bar", "foo", "baz"}))
-	fmt.Printf("%#v\n", UniqueStrings([]string{"foo", "bar", "baz"}))
-	fmt.Printf("%#v\n", UniqueStrings(nil))
-	// Output:
-	// []string{"foo", "bar", "baz"}
-	// []string{"foo", "bar", "baz"}
-	// []string{}
-}
-
 func ExampleIsEqualStringSlice() {
 	s1 := []string{"foo", "bar", "baz"}
 	s2 := []string{"foo", "bar", "baz"}
