+++ slice/min_max_avg_test.go
@@ -1,0 +1,1 @@
+package slice
